% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic_taxa.R
\name{classic_taxa}
\alias{classic_taxa}
\title{Utility for compositional microbiota}
\usage{
classic_taxa(
  physeq,
  taxrank,
  groupvar,
  perc_threshold = 1,
  Abundance = Abundance
)
}
\arguments{
\item{physeq}{Phyloseq object with relative abundance in feature table}

\item{taxrank}{Taxonomic rank to simplify}

\item{groupvar}{Experimental grouping variable}

\item{perc_threshold}{Percentage of relative abundance at which taxa will be considered other}

\item{Abundance}{Name of the column where Abundance is stored}
}
\value{
Melted data frame with taxa filtered by relative abundance threshold
}
\description{
This function generates a data frame for relative abundance plots
}
