% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESEQ2Groups.R
\name{DESEQGroups}
\alias{DESEQGroups}
\title{Performs DESeq analysis for microbiota data using two groups}
\usage{
DESEQGroups(physeq, groupvar, alpha = 0.05, lfc_threshold = 5)
}
\arguments{
\item{physeq}{Phyloseq object containing count data}

\item{groupvar}{Grouping variable}

\item{alpha}{Significance level}

\item{lfc_threshold}{Log Fold Change cutoff for biological relevance}
}
\description{
Performs DESeq analysis for microbiota data using two groups
}
