% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvolcano.R
\name{ggvolcano}
\alias{ggvolcano}
\title{Title}
\usage{
ggvolcano(df, xval = log2FoldChange, yval = padj, group = group)
}
\arguments{
\item{df}{Data frame containing differential expression results}

\item{xval}{Value for x axis (Effect Size)}

\item{yval}{Value for y axis (p adjust)}

\item{group}{Group Variable}
}
\description{
Title
}
