% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_qiime.R
\name{load_from_qiime}
\alias{load_from_qiime}
\title{Loads qiime files into R}
\usage{
load_from_qiime(
  dir = "qiime",
  metadata_ext = ".txt",
  features = "table",
  taxonomy = "taxonomy",
  metadata = "metadata",
  tree = "rooted-tree"
)
}
\arguments{
\item{dir}{Directory where qiime generated files will be present}

\item{metadata_ext}{Extension for metadata file}

\item{features}{Feature table name}

\item{taxonomy}{Taxonomy table name}

\item{metadata}{Metadata file name}

\item{tree}{Tree file name, using rooted-tree by default}
}
\value{
A phyloseq object
}
\description{
depends on qiime2R
}
