% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculates mean ± standard deviation}
\usage{
mean_sd(z)
}
\arguments{
\item{z}{Column}
}
\value{
mean ± standard deviation
}
\description{
Used in conjunction with dplyr::summarise_all
}
