% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_statistic.R
\name{multi_statistic}
\alias{multi_statistic}
\title{Calculates ANOVA o KW depending on normality test}
\usage{
multi_statistic(df, group_name, physeq = data, alpha = 0.05)
}
\arguments{
\item{df}{Data frame with parameter to be evaluated}

\item{group_name}{Name of the experimental groups}

\item{physeq}{Phyloseq object containing sample data}

\item{alpha}{Siginifcance level}
}
\description{
Calculates ANOVA o KW depending on normality test
}
