% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compositional.R
\name{plot_compositional}
\alias{plot_compositional}
\title{Plot compositional microbiota from data frame}
\usage{
plot_compositional(df, taxrank, groupvar, Abundance = Abundance)
}
\arguments{
\item{df}{Relative Abundance data frame with features filtered}

\item{taxrank}{Taxonomic rank to plot}

\item{groupvar}{Experimental Group}

\item{Abundance}{Column were the relative abundance is stored}
}
\description{
Plot compositional microbiota from data frame
}
