% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Put mean, sd and hypothesis test in a data frame}
\usage{
summary_stats(df, group_var, physeq = data)
}
\arguments{
\item{df}{Data frame}

\item{group_var}{Experimental groups}

\item{physeq}{Phyloseq object}
}
\description{
Put mean, sd and hypothesis test in a data frame
}
